% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biacore.R
\name{plot_biacore}
\alias{plot_biacore}
\title{Plot biacore SPR data}
\usage{
plot_biacore(file, raw_color = "black", fitted_color = "red")
}
\arguments{
\item{file}{A txt file containing SPR data from biacore T200}

\item{raw_color}{A character string containing color information in hex for
the raw data}

\item{fitted_color}{A character string containing color information in hex for
the fitted data}
}
\value{
A ggplot2 object
}
\description{
Plot biacore SPR data
}
\examples{
file <- system.file("extdata", "aCD16_3G8_Mouse.txt", package = "biacore")
plot_biacore(file)
}
